requirements.txt
torch
torchvision
numpy
pandas
opencv-python
scikit-learn
matplotlib
umap-learn
streamlit
geopandas
folium
streamlit-folium
pyyaml
pillow
tqdm
scipy
plotly
